[SR][Locked Figures] Give all locked figures "img" role
Right now, when traversing through locked figures with a
screen reader, it reads out the label and then identifies
the figure as "empty group." This is an extremely
unhelpful description.

To give a more accurate description of the function the
locked figures are serving, we need to give them a
role of "img." This way, the screen reader will read
the aria label and then say "image" after.

Issue: none

Test plan:
Storybook
- http://localhost:6006/iframe.html?globals=&args=&id=perseuseditor-widgets-interactive-graph--mafs-with-locked-figures&viewMode=story
- Go through this story before and after the change. It should say image
  isntead of empty group now.