[benc/no-synthetic-undefineds] Don't add undefined values to objects when parsing JSON
Previously, when parsing typed objects from JSON, we would add `[key]:
undefined` for any fields that were specified in the parser schema but
were absent from the JSON. This was a difference from the behavior of
`JSON.parse()` that could cause bugs in some cases (e.g. if a parsed
object was spread into another object, it could overwrite existing
values with `undefined` when that was not intended).

Thus, to better match the behavior of `JSON.parse()`, we now omit
`undefined` values from the parsed object if they are not present on the
original object.

Issue: LEMS-2774

Test plan:

`yarn test`