[Bugfix] Zoomable content within a collapsed parent not showing when expanded (Mobile only) (#2185)
## Summary:
Content that is intended to be zoomable needs to have a context for its available space. It does this by using the `offsetWidth` of its container node to calculate a `scale` value. However, if the container is within a collapsible widget (such as the Explanation widget), then the container node has a zero width, which results in a scale of zero making it unavailable to the learner.

This bugfix adds a check for the `offsetWidth` value and loops up the DOM tree for an ancestor with the `perseus-renderer` class. This ancestor should have a valid width. If no ancestors can be found with a valid width (aka we run out of `perseus-renderer` nodes), then we break out of the loop and allow for the scale to be zero (instead of causing an infinite loop or error message).

Issue: LEMS-2805

## Test plan:

1. Go to the Test Everything Course (Settings > Debug Settings).
2. Go to Unit 2: > Radio > “Radio, definition, explanation exercise”.
3. Answer a question correctly.
4. Click the link in the popup (e.g. “See how we answered this”) to view the hint/rationale.
   * (If needed) Scroll up to the Explanation widget.
5. Toggle open the Explanation widget.
6. Observe that the open widget renders the content in a scaled-down mode (so that there is no scrolling).
   * Observe that the content can be scaled up by tapping once on the content.

Author: mark-fitzgerald

Reviewers: jeremywiebe, mark-fitzgerald

Required Reviewers:

Approved By: jeremywiebe

Checks: ❌ Publish npm snapshot (ubuntu-latest, 20.x), ✅ Cypress (ubuntu-latest, 20.x), ✅ Check for .changeset entries for all changed files (ubuntu-latest, 20.x), ✅ Check builds for changes in size (ubuntu-latest, 20.x), ✅ Lint, Typecheck, Format, and Test (ubuntu-latest, 20.x), ✅ Publish Storybook to Chromatic (ubuntu-latest, 20.x)

Pull Request URL: https://github.com/Khan/perseus/pull/2185