[unlimited-focus] [Bug Fix] Ensure that Unlimited Graphs move focus properly when deleting points out of order.
This PR is part of the Interactive Graph Project.

While recreating LEMS-2551, it was discovered that focus was not being set for mouse-based interaction. This PR fixes both that issue, and the issue from LEMS-2551 where deleting the points out of order would eventually result in a failure to move the focus to the next point.

Issue: LEMS-2551

Test plan:
tests pass