[get-rid-of-disgusting-function-call] [Interactive Graph] Replace string-based function call with switch statement
There's a part in the interactive graph code where the equation string
getter function is called by a code that concatonates a string together
and uses that string as a function name to call.

For the sake of type safety and the wellbeing of our engineers, updating
this to a switch statement.

Issue: none

Test plan:
`yarn jest packages/perseus-editor/src/widgets/__tests__/interactive-graph-editor.test.tsx`

Storybook
- http://localhost:6006/?path=/story/perseuseditor-editorpage--demo