[update-protractor] [Interactive Graph] Update protractor to new SVG
The old protractor looked awful. We got a request to use a protractor
with two rows of numbers going in opposite directions (previously
there was only one direction).

- Update the protractor to a new SVG with numbers going both ways

Note: The numbers are kind of hard to read over the axes. I tried to add
white backgrounds to the numbers, but it's unfortunately not as easy as
it looks. Since this is not a regression, I left it as is.

Note 2: There is a noticable performance hit with this implementation.
If someone has ideas for how to improve the performance, please let me know!

Issue: https://khanacademy.atlassian.net/browse/LEMS-2863

Test plan:
Storybook
- https://khan.github.io/perseus/?path=/story/perseus-widgets-interactive-graph-visual-regression-tests--mafs-with-protractor