[LEMS-2849] Answerless Radio (#2233)
## Summary:
So this focuses on `numCorrect` which originally derived from the answers but now is its own widget option. I'm sure there's an argument that this also shouldn't be sent to the FE, but it's conditionally used to render help text: `Select 2 answers`. I don't know, I don't think it's that big of a deal to have it but I also made it optional so we can remove it from questions that don't need it: `needsNumCorrect = options.multipleSelect === true && options.countChoices === true`.

The important thing is that we can render, answer, and score a Radio widget that's been stripped of answers. As far as I can tell, this supports all behavior pre-scoring.

Issue: LEMS-2849

## Test plan:
Radio should continue to be renderable, answerable, and scorable.

Author: handeyeco

Reviewers: handeyeco, benchristel, jeremywiebe, Myranae

Required Reviewers:

Approved By: benchristel, jeremywiebe

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2233