[LEMS-2852] Answerless Expression (#2226)
## Summary:
So the main thing with the Expression widget is that it was using answer data to determine which `extraKeys` to show in the `MathInput` keypad. For instance if a possible answer was `42i`, it would show `i` in the `Extra` tab on the keypad.

This PR sets us up for removing answers by adding `extraKeys` to the data schema so it can be determined at publish time rather than read time. We simulate this by upgrading the widget to provide `extraKeys` when not present on the data.

Issue: LEMS-2852

## Test plan:
- Go to an Expression widget that has a constant/variable in the answer
- It should render the same, including having the constant/variable in the keypad
- It should be answerable/scorable

Author: handeyeco

Reviewers: handeyeco, jeremywiebe

Required Reviewers:

Approved By: jeremywiebe

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2226