Add a basic FreeResponse widget (#2273)
## Summary:
This commit adds a basic version of the new FreeResponse widget. This
widget allows users to enter free text into an input field. The initial
use case is for "short answer" type questions.

The UX of the widget is not complete in this commit. Instead, the
intention is to get all of the files in place with a basic UI which will
then be iterated on with more widget options and more sophisticated
rendering.

Issue: https://khanacademy.atlassian.net/browse/LIT-1661

## Test plan:
- View the editor and rendered widget stories in Storybook.
- Run the tests.

Author: aag

Reviewers: benchristel, aag, handeyeco, rgpass

Required Reviewers:

Approved By: handeyeco

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2273