[SR] Add aria-hidden to individual components of interactive graph elements (#2277)
## Summary:
When using Safari + VoiceOver, the screen reader reads out every single internal part of every single element on the graph. For example, the arrow on a line is read out as "image" as well as the segment composing the line. This means that when traversing the graph in Safari, there are a whole bunch of empty "image"s being read out with no context. The context is provided by specific element's aria labels.

To remedy this, I added `aria-hidden` to all the visual elements that should not be read as "image" in all the graph types.

Issue: none

## Test Plan:
Storybook
- http://localhost:6006/?path=/docs/perseuseditor-widgets-interactive-graph--docs
- Go to every graph type and traverse through it with a screen reader
- Confirm there are no random empty "image" calls

Author: nishasy

Reviewers: SonicScrewdriver, catandthemachines, mark-fitzgerald

Required Reviewers:

Approved By: SonicScrewdriver, catandthemachines

Checks: ✅ 13 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2277