[poly-gets-updated-coords] docs(changeset): Bugfix to ensure that the Interactive Graph Editor updates the Polygon start coords if the number of sides change. (#2283)
## Summary:
This PR is part of the Interactive Graph project. 

The purpose of this simple PR is to implement a fix that ensures that the start coords are always updated when content creators update the number of sides for the graph. 

Issue: LEMS-2883

## Test plan:
- manual testing

Author: SonicScrewdriver

Reviewers: nishasy

Required Reviewers:

Approved By: nishasy

Checks: ✅ 8 checks were successful, ⏹️  5 checks were cancelled

Pull Request URL: https://github.com/Khan/perseus/pull/2283