[its-a-drag-to-tap-out] docs(changeset): Ensure that dragging a point also moves the focus for Interactive Graph (#2282)
## Summary:
This PR is part of the Interactive Graph project. 

The purpose of the PR is to ensure that dragging a point with a touch event also moves the focus to that point. This is to ensure that the focus always remains on the last element interacted with, which avoids confusion when the user tries to delete points or move them with the keyboard. 

NOTE: I wasn't able to fix a bug related to the hover state sizing of the control points, as I simply cannot seem to find where this is occurring. As a result, touch-dragging a point without any delay can result in the point not getting "bigger". However, the focus outline, hairlines, and focus itself all move to the correct element. I've attached a video below to demonstrate this.

## Video:
https://github.com/user-attachments/assets/59e26360-b54e-4340-b49b-c4f9c32c4baf

Issue: LEMS-2475

## Test plan:
- Manual testing

Author: SonicScrewdriver

Reviewers: SonicScrewdriver, nishasy

Required Reviewers:

Approved By: nishasy

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2282