[dial-back] Ensure that keypad backspaces don't delete entire parentheses nodes at once.
When performing a backspace, our Math Input keypad appears to be deleting the entire contents of a parentheses node, which has been frustrating for our users.

For example: hitting backspace on x+(9*9) would result in only x+ remaining.

When reviewing the code it appears that this functionality was set up deliberately, but I have been unable to find the justification or history why.

Given there's no clear reason for this feature, this PR is removing this special logic so that backspaces are handled with the default logic.

For example: hitting backspace on x+(9*9) would result in x+(9*9

Issue: MOB-5576

Test plan:
Manual testing