[update-axes-labels-sr] [SR] Move axis labels' screen reader text
The current implementation has it so that the label "X-axis"/"Y-axis"
is on the group surrounding the TeX that renders the axis label. This
is a little awkard and confusing to understand.

We also cannot put the TeX content directly in the aria-label because
that won't read the math correctly.

Instead, I've put added "X-axis" and "Y-axis" as sr-only text within
the group, and I've removed them from the group labels. This way, the
screen reader can read "X-axis" directly followed by the axis label,
rather than adding info about the group containing the axis label.

Issue: none

Test plan:
- Go to http://localhost:6006/iframe.html?globals=&args=&id=perseuseditor-widgets-interactive-graph--interactive-graph-segment&viewMode=story
- Use a screen read to go through the graph
- Confirm that going through with a screen reader, it says "X-axis"
  followed by "x, math" (same for Y-axis)
- Change axis labels in the editor
- Confirm that it reads the updated label correctly as well
- Repeat for a math label within `$...$` block