Don't export mediaQueries - not used (and shouldn't be) outside of Perseus (#2373)
## Summary:

I noticed in webapp that the Perseus `mediaQueries` object is being used. It was unrelated to Perseus and so I fixed it. But it pointed to the fact that there's no reason for this object to be exported. 

I checked other packages in this repo and none of them use `mediaQueries`, so I'm changing things so that we don't export it anymore; it's an internal thing.

Issue: "none"

## Test plan:

`pnpm typecheck`
Check consuming apps and ensure none use `mediaQueries`

Author: jeremywiebe

Reviewers: mark-fitzgerald

Required Reviewers:

Approved By: mark-fitzgerald

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2373