When parsing Perseus JSON, optionalize object properties that can be undefined (#2397)
## Summary:
Previously, it was not possible for the return types of our parsers to have
optional properties, like `{foo?: string}`. This meant that the data-schema types
that used that syntax were not mutually assignable with the types inferred from their parser.

E.g. if you have a parser like this:

```ts
const foo = object({foo: optional(string)});
```

The type `ParsedValue<typeof foo>` is equivalent to `{foo: string |
undefined}`. It is assignable to `{foo?: string}`, but `{foo?: string}`
is not assignable to `ParsedValue<typeof foo>`.

This lack of mutual assignability made it hard to write robust typetests for
the parsers, which in turn made it hard to ensure that parsers would be kept
up to date when changes were made to data-schema.

For example, consider the case of adding a variant to a union type in data-schema.

```ts
type MyUnion = "a" | "b" | "new-variant";
```

Suppose the parser for this type looks like this:

```ts
const parseMyUnion = enumeration("a", "b");
```

You won't get any type errors after adding `"new-variant"` to the type, because
`"a" | "b"` is assignable to `"a" | "b" | "new-variant"`. Yet the parser will
reject `"new-variant"`, which is certainly not what you want. In order to catch
this bug statically, you need typetests that ensure the data-schema types can
be assigned to the parser return types.

This PR takes a first step toward making that possible, by standardizing on the
syntax `{foo?: T}` for optional properties, rather than `{foo: T | undefined}`.

Issue: LEMS-3048

## Test plan:

`pnpm tsc`

Author: benchristel

Reviewers: benchristel, jeremywiebe, handeyeco, mark-fitzgerald, Myranae, SonicScrewdriver

Required Reviewers:

Approved By: jeremywiebe

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2397