[add_formatting_rendering] Render LaTeX in Free Response Widget questions
This commit adds code to render LaTeX formatted questions in the
Free Response Widget. The specific use case we want to support is
including Math expressions in the question. We use the same Renderer
that is used to render LaTeX elsewhere, but we don't want to support
nested widgets in the question, so we detect that as an issue when
saving in the editor.

Note that the spacing around the question is changed in this commit and
it doesn't match the designs, but that will be fixed in a later change.

Issue: https://khanacademy.atlassian.net/browse/AX-964

Test plan:
- View the Free Response Widget in Storybook and verify that the
  "Question with Katex" story renders properly.