Use interface for shared ref type (#2452)
## Summary:

The `numeric-input` widget uses two different input components depending on whether a custom keypad is enabled or not. We want to retain a ref to either of these inputs within the widget. Originally, the ref was defined as the union of these two components, but that leads to sporadic type check failures. 

This PR adjusts things so we define a common `interface` for these two components and then use that interface as the type for hte ref. This makes it so the ref is a very narrow type referencing only the fields that are used.

Issue: "none"

## Test plan:

`yarn typecheck`
`yarn tsc --watch` (make some changes, save, then revert to ensure the type checking doesn't fail)

Author: jeremywiebe

Reviewers: benchristel, jeremywiebe, SonicScrewdriver, handeyeco

Required Reviewers:

Approved By: benchristel

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2452