Render TeX in Free Response Widget questions (#2453)
This commit adds code to render TeX formatted questions in the
Free Response Widget. The specific use case we want to support is
including Math expressions in the question. We use the same Renderer
that is used to render TeX elsewhere, but we don't want to support
nested widgets in the question, so we detect that as an issue when
saving in the editor.

Note that the spacing around the question is changed in this commit and
it doesn't match the designs, but that will be fixed in a later change.

Issue: https://khanacademy.atlassian.net/browse/AX-964

- View the Free Response Widget in Storybook and verify that the
  "Question with Tex content" story renders properly.

![image](https://github.com/user-attachments/assets/e3275716-4d0b-4100-b2c1-bd25072c8e3d)

Author: aag

Reviewers: jeremywiebe, handeyeco

Required Reviewers:

Approved By: jeremywiebe

Checks: ✅ 13 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2453