[free_response_styling_a11y] Update Free Response styles and accessibility
This commit updates the styles and layout of the Free Response Widget to
match the designs. It also updates some markup to improve the
accessibility of the widget, including using the `LabeledField` Wonder
Blocks component.

The designs can be found in Figma:
https://www.figma.com/design/h5nZthMLTd56M0GLYsrDry/ELA---Free-Response?node-id=10184-1509&t=7etutvzW6ikoM9cS-1

Issue: https://khanacademy.atlassian.net/browse/AX-963

Test plan:
- View the widget variants in Storybook.
- Verify that the rendered widgets match the designs.