Add registry guard (#2483)
## Summary:
[See this thread](https://khanacademy.slack.com/archives/C08B8MLKPUZ/p1747236599339659?thread_ts=1747179744.422339&cid=C08B8MLKPUZ)

The root of the problem we're trying to solve here:

- We rely on a runtime registry of widgets; right now we do this for `perseus` (both regular registry and editor registry), `perseus-core`, and `perseus-score` (`perseus-core` and `perseus-score` are likely to get merged).
- These registries map a component type to a chunk of logic specific to the widget.
- If we try to access a widget that's not registered, we just silently fail (maybe we should throw an error though, why are we trying to access widgets that don't exist?)
- If we forget to call the register function before accessing the registry, we silently fail for all widgets.

Proposed solution:

It's doesn't really make sense for a registry to be empty; it almost certainly means something is wrong. So this proposed solution is to initialize the registry using a dummy object that immediately throws whenever it's accessed. The first time we register a widget, we replace the dummy object with a real object. This ensures that at least one widget is registered; since we usually register widgets altogether, it makes it much more likely that we're accessing the registry in a state we're expecting.

Alternative solution:

We could just make all the registry accessor functions throw if they don't find a widget they were expecting to find. We could allow the property we're looking for to have a default, but if the widget type _doesn't exist at all_ maybe we could throw? That would spread out a lot of logic and require new functions to do the same checks; it could also mean we're registering widgets we don't want to register just to get past the check (like unscored widgets in `perseus-score`).

I pitched this draft PR to both @benchristel and @jeremywiebe. Ben said specifically that he doesn't like the proxy. Jeremy said specifically that he does like the proxy. Begin battle!

Author: handeyeco

Reviewers: handeyeco, benchristel, jeremywiebe

Required Reviewers:

Approved By: benchristel

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2483