docs(changeset): New feature for PhET Widget to provide an option for full screen on mobile devices. (#2440)
## Summary:
This PR creates a fake fullscreen view in the mobile application for the PhET Widget, as we're currently unable to use the fullscreen API. It also adds a new apiOption for isMobileApp, which will allow Perseus to know when we're being served in one of the mobile applications. 

This PR will require updates in webapp in order to pass down the new isMobileApp variable as part of the apiOptions.

Additional support for the fullscreen API could _**possibly**_ be added to the Mobile application in the future, but there is currently no set release date or work for the next Mobile release. Given that, we've opted for a bit of a "hacky" temporary solution that will allow us to get the widget a little larger to improve usability. 

Unfortunately, full screen will not be supported in iOS web at this time. 

This direction was chosen due to several restrictions:

1. We're locked into a z-index in an upstream container that we cannot change. 
2. We have many variable headers, banners, and footers in the mobile web version that interfere with allowing this fix to also be served to mobile web. 
3. We can't inject the widget at the document root to work around the z-index issue as it would result in the loss of any actions that had been performed in the PhET iframe. 
4. This is a low usage widget. If we were to see increased usage and partnership, we could discuss improving the communication of state with PhET so that alternative solutions could be used in the future. 

Issue: LEMS-2922

## Test plan:
- Manual testing in mobile application 
- New test that confirms isMobileApp logic is being used.

Author: SonicScrewdriver

Reviewers: SonicScrewdriver, jeremywiebe, mark-fitzgerald

Required Reviewers:

Approved By: jeremywiebe, mark-fitzgerald

Checks: ✅ 13 checks were successful, ⏹️  5 checks were cancelled

Pull Request URL: https://github.com/Khan/perseus/pull/2440