Remove preferred popover direction from Label Image editor (#2512)
## Summary:
The option from the preferred pop-over direction dropdown was not persisting into the JSON for the exercise, and it did not show the chosen option to the content author. This removes the broken dropdown.

### Before
<img width="500" alt="Label Editor with preferred popover direction" src="https://github.com/user-attachments/assets/130ead8d-a4a0-48e5-b452-2015f8992938" />


### After
<img width="500" alt="Label Editor without preferred popover direction" src="https://github.com/user-attachments/assets/2233c653-75c0-48bb-a6ba-396a765a686e" />

There is potential future work here to re-implement this functionality in the editor and with Label Image if possible. See LEMS-3126.

Issue: LEMS-3110

## Test plan:
- Confirm all tests and checks pass
- Confirm the label image editor works as expected via Storybook and ZND
- Confirm the PreferredPopoverDirection dropdown is gone.

Author: Myranae

Reviewers: handeyeco, benchristel

Required Reviewers:

Approved By: handeyeco, benchristel

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2512