[fix_get_save_warnings] Fix getSaveWarnings in Free Response widget
This commit changes the getSaveWarnings() function definition in the
Free Response widget to be defined using public class field syntax of
the public method syntax. Previously, `this` wasn't being bound
properly, so the function was unable to read `this.props`.

The commit also adds some more type annotations and changes all methods
on the widget to be defined in the same way.

Issue: https://khanacademy.atlassian.net/browse/AX-1005

Test plan:
- Load the PR's Perseus bundle in local webapp.
- Add a Free Response widget in the exercise editor and verify that it
  renders properly.