Bugfix: Answerless -> Answerful clearing widget state (#2479)
## Summary:
So basically when we go from one question to the next question, we want to clear the widget state. That makes sense; we don't want a learner's selection to move forward to a totally different question. But what does it mean to have different questions? Two questions could have the same content (ie `solve for x [[☃ expression 1]]`). It might be unlikely, but we could go between two exercises and land on the same `problemNum`...so what do we do?

The original logic was doing a deep equality check on `widgets` which makes sense. It's _really_ unlikely that we'll go between two exercises and get (1) the same content (2) the same problem number _and_ (3) the same exact widget options. If any of those changed, the original code wiped widget state clean.

Unfortunately with SSS, we now want to switch between answerful and answerless data. The `content` will be the same, `problemNum` will be the same, but the data in `widgets` will change and we _don't_ want to clear the state.

So this proposed fix does a comparison between prev/next props but only looks at the answerless data to see if we're on a new question.

Issue: LEMS-2948

## Test plan:
- Go to a story that starts answerless; [like this one](https://khan.github.io/perseus/?path=/story/perseus-widgets-radio--answerless-single-select)
- Answer the question
- Click the "Check" button
- Under the hood, the data should switch from answerless to answerful
- Expected:
  - Before: the user input would be cleared
  - Now: the user input will not be cleared

Author: handeyeco

Reviewers: benchristel, handeyeco, Myranae

Required Reviewers:

Approved By: Myranae

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2479