Add support for CSS Modules (#2431)
## Summary:
In our effort to modernize our approach to and toolset for styling, this change adds the ability to use standard CSS syntax in a `*.module.css` file for any of our widgets and renderers. This will enable us to migrate our styling away from Aphrodite and onto a standardized format. This is accomplished via an update to our build process. The prior Rollup `styles` plugin has been replaced with the `postcss` plugin since the [LESS files have since been converted to standard CSS](https://github.com/Khan/perseus/pull/2489) (no "building" is necessary).

The conversion of Aphrodite code to standard CSS is accomplished by the addition of a script that can be run against any TSX file that contains Aphrodite code. This script file uses various `@babel` plugins to generate an AST parsing of the code files, and then procedurally extracts the CSS settings and reformats them into a separate CSS file. This script does the heavy-lifting of moving and reformatting the styling, but it is expected that there are nuances in the JavaScript code that will necessitate adjustments to complete the transformation. For instance, in places where style information was passed to a `style` attribute, it will now need to reference the generated class name in the `className` attribute. This script is intended to be temporary, and only last as long as needed to convert and remove Aphrodite.

Issue: LEMS-3008

## Test plan:
1. Review the `<fieldset>` element in the Radio widget ([line 211 in radio/base-radio.new.tsx](https://github.com/Khan/perseus/pull/2431/files#diff-1e7272f36753eeee0138a5375f4434db42e6701e75782f0ab1638b99a53d2602R211)).
   * Note the addition of a class (`cssModulesTest`) pulled from a CSS file (radio/radio-test.module.css)
1. Launch Storybook
1. Navigate to the [new Radio Widget](http://localhost:6006/?path=/story/perseus-widgets-radionew--single-select) (this only shows when the Feature Flag is on)
1. Inspect the instructions text "Choose 1 answer:"
   * Note that the `<fieldset>` parent element in the DOM now includes a class called something like `_css-modules-test_6k65e_1` (<underscore><my-class-name><underscore><8-digit-hash>)
   * Note that in PROD, the class name will be `perseus_<8-digit-hash>`

Author: mark-fitzgerald

Reviewers: mark-fitzgerald, ivyolamit, jeremywiebe, SonicScrewdriver, anakaren-rojas

Required Reviewers:

Approved By: ivyolamit, jeremywiebe

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2431