[ig-editor-stop-autoupdate] [CX] Stop steps from auto-updating when range is changed in Interactive Graph editor
Right now, when a content author changes the range in the interactive graph settings,
the tick step, grid step, and snap step all update to a specific ratio of the range.

Sometimes content authors don't want this to happen. We've received a request to
stop the auto-updates.

In this PR, I'm stopping the auto-updates, and adding a button that can be clicked
if a content author wants the steps updated the way they used to be.

Issue: https://khanacademy.atlassian.net/browse/LEMS-3217

Test plan:
`pnpm jest packages/perseus-editor/src/widgets/interactive-graph-editor/components/interactive-graph-settings.test.tsx`

Storybook
- Go to http://localhost:6006/iframe.html?globals=&args=&id=perseuseditor-widgets-interactive-graph--interactive-graph-with-aria-label&viewMode=story
- Change the x Range min to -100
- Confirm that the preview graph shows an error instead
- Press the "Auto-adjust steps" button
- Confirm that the graph shows up normally again
- Confirm that the tick step, grid step, and snap step are all updated
- Repeat for x Range max, y Range min, and y Range max