Radio: renderer user input state (#2676)
## Radio Summary:

I waited to do Radio last (besides the Group widgets) because it's a tricky one. Radio's UserInput is tricky because it gets shuffled and unshuffled, Radio shoves a lot of state in Renderer when it really doesn't seem to need to, and on top of Radio itself being tricky, there's a lot of active development on Radio that makes it a kind of moving target.

Ideally I wanted to wait until Radio's UserInput refactor but I _**have**_ to move on to the Group widgets (this work has already been a month-long unexpected leg of work) and I have to get Radio done before I can move on to those.

This PR is hopefully an acceptable compromise between the Radio team and SSS. It's a relatively shallow change that just adapts the data that the Renderer needs to the data that the Radio team needs. Most of the changes are adding or cleaning up tests, adding comments, and moving some code to conform to the new expectations from [LEMS-2566](https://github.com/Khan/perseus/pull/2566).

## Generic Summary:

See the [parent PR](https://github.com/Khan/perseus/pull/2566) for more context. Part of [LEMS-3208](https://khanacademy.atlassian.net/browse/LEMS-3208).

In #2566 I add additional APIs to let widgets consume `userInput` from a parent and to update that parent state using `handleUserInput`, while also supporting the legacy way of storing user input (which was a combination of internal widget state and stashing state in a random blob in Renderer). This PR is part of a series of PRs that go widget-by-widget to use the new API.

This is a generic summary I'm putting on each PR and not every point will be applicable to every widget, but the common themes are:

- I added a test for serialization to make sure we're still backwards compatible (see [LEMS-3185](https://khanacademy.atlassian.net/browse/LEMS-3185)), then implemented the helpers `getSerializedState` and `getUserInputFromSerializedState` to keep supporting the expected results.
- I moved user input for the component into the new `userInput` state in Renderer. This means consuming `userInput` in the component and calling `handleUserInput` on change.
- When `transform` did something to initialize user input state, I moved the logic to `getStartUserInput`.
- When `staticTransform` did something to get correct state in static widgets, I moved the logic to `getCorrectUserInput`.
- Editors that use a widget to collect answer data are changed to support the new API.

Please see the [parent PR](https://github.com/Khan/perseus/pull/2566) for more information.

[LEMS-2566]: https://khanacademy.atlassian.net/browse/LEMS-2566?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[LEMS-3208]: https://khanacademy.atlassian.net/browse/LEMS-3208?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Author: handeyeco

Reviewers: handeyeco, benchristel, SonicScrewdriver, ivyolamit, jeremywiebe, Myranae, mark-fitzgerald, anakaren-rojas, nishasy

Required Reviewers:

Approved By: SonicScrewdriver, ivyolamit

Checks: ✅ 8 checks were successful

Pull Request URL: https://github.com/Khan/perseus/pull/2676